first batch of changes
