+++ ingester.Dockerfile
@@ -42,7 +42,7 @@ COPY data/catalogs/ /app/data/catalogs/
 COPY data/ztf_matchfiles/ /app/data/ztf_matchfiles/
 COPY data/ztf_source_features/ /app/data/ztf_source_features/
 COPY data/ztf_source_classifications/ /app/data/ztf_source_classifications/
-COPY data/TURBO_alerts/ /app/data/TURBO_alerts/
+COPY data/turbo_alerts/ /app/data/turbo_alerts/
 
 
 # copy over the config and the code
@@ -51,13 +51,13 @@ COPY ["config.yaml", "version.txt", "kowalski/generate_supervisord_conf.py", "ko
       "kowalski/alert_broker.py",\
       "kowalski/alert_broker_ztf.py",\
       "kowalski/alert_broker_winter.py",\
-      "kowalski/alert_broker_TURBO.py",\
+      "kowalski/alert_broker_turbo.py",\
       "kowalski/ops_watcher_ztf.py",\
       "kowalski/tns_watcher.py",\
       "kowalski/alert_broker_pgir.py",\
       "kowalski/dask_cluster_pgir.py",\
       "kowalski/dask_cluster_winter.py",\
-      "kowalski/dask_cluster_TURBO.py",\
+      "kowalski/dask_cluster_turbo.py",\
       "kowalski/ingester.py",\
       "kowalski/performance_reporter.py",\
       "kowalski/requirements_ingester.txt",\
@@ -65,11 +65,11 @@ COPY ["config.yaml", "version.txt", "kowalski/generate_supervisord_conf.py", "ko
       "tests/test_alert_broker_ztf.py",\
       "tests/test_alert_broker_pgir.py",\
       "tests/test_alert_broker_wntr.py",\
-      "tests/test_alert_broker_TURBO.py",\
+      "tests/test_alert_broker_turbo.py",\
       "tests/test_ingester_ztf.py",\
       "tests/test_ingester_pgir.py",\
       "tests/test_ingester_wntr.py",\
-      "tests/test_ingester_TURBO.py",\
+      "tests/test_ingester_turbo.py",\
       "tests/test_tns_watcher.py",\
       "tests/test_tools.py",\
       "tools/fetch_ztf_matchfiles.py",\
