+++ kowalski.py
@@ -478,13 +478,7 @@ class Kowalski:
             {
                 "part": "TURBO alert broker components",
                 "container": "kowalski_ingester_1",
-                "test_script": "test_alert_broker_TURBO.py",
-                "flaky": False,
-            },
-            {
-                "part": "TURBO alert ingestion",
-                "container": "kowalski_ingester_1",
-                "test_script": "test_ingester_TURBO.py",
+                "test_script": "test_alert_broker_turbo.py",
                 "flaky": False,
             },
             {
@@ -503,6 +497,12 @@ class Kowalski:
                 "part": "WINTER alert broker components",
                 "container": "kowalski_ingester_1",
                 "test_script": "test_alert_broker_wntr.py",
+                "flaky": False,
+            },
+            {
+                "part": "TURBO alert ingestion",
+                "container": "kowalski_ingester_1",
+                "test_script": "test_ingester_turbo.py",
                 "flaky": False,
             },
             {
